+++ src/components/Projects/Projects.js
@@ -1,36 +1,9 @@
-import React, { useState } from 'react';
 import { Box, Typography } from '@mui/material';
 import AsyncHelmet from '../AsyncHelmet';
-import ProjectCards from './ProjectCards';
 import projectData from '../../data/projects.json';
 import FilteringMenu from './FilteringMenu';
-import { useFilter } from './useFilter';
 
 const Projects = () => {
-  const [selectedChips, setSelectedChips] = useState(new Set());
-  const [searchText, setSearchText] = useState('');
-
-  const { filterByUniqueProjects, filterBySelectedChips } = useFilter();
-
-  // Sorts all projects chronologically by start date
-  const sortedProjectData = [...projectData].sort((a, b) => {
-    return new Date(b.dateStarted) - new Date(a.dateStarted);
-  });
-
-  // Extract project names
-  const projectNames = projectData.map((project) => project.title);
-
-  // Extracts the unique tools used across all projects
-  const uniqueTools = filterByUniqueProjects({ sortedProjectData });
-
-  // Filter projects based on selected chips
-  const filteredProjects = filterBySelectedChips({ selectedChips, sortedProjectData });
-
-  // Further filter projects based on search text
-  const searchedProjects = filteredProjects.filter((project) =>
-    project.title.toLowerCase().includes(searchText.toLowerCase())
-  );
-
   return (
     <Box
       id="projects"
@@ -47,16 +20,7 @@ const Projects = () => {
         Projects
       </Typography>
 
-      <FilteringMenu
-        projectNames={projectNames}
-        searchText={searchText}
-        setSearchText={setSearchText}
-        uniqueTools={uniqueTools}
-        selectedChips={selectedChips}
-        setSelectedChips={setSelectedChips}
-      />
-
-      <ProjectCards searchedProjects={searchedProjects} />
+      <FilteringMenu projectData={projectData} />
     </Box>
   );
 };
